another views naming fix (#37)
* pro gamer move once more

* done

* *test commit*

* undoing test commit

* components upd
