--- Buildlease/Buildlease/ClientApp/src/components/Views/CategoryFilterView.tsx
@@ -1,10 +1,0 @@
-export default interface CategoryFilterView {
-  Id: number,
-  Name: string,
-
-  UnitName: string | null,
-  MinValue: number | null,
-  MaxValue: number | null,
-  
-  Values: string[] | null,
-}
