--- Buildlease/Buildlease/ClientApp/src/components/Views/CategoryFullView.tsx
@@ -1,7 +1,0 @@
-export default interface CategoryFullView {
-  Id: number,
-  ParentId: number,
-  Name: string,
-  ProductCount: number,
-  DefaultImagePath: string,
-}
