--- Buildlease/Buildlease/ClientApp/src/components/Views/CategoryView.tsx
@@ -1,4 +1,0 @@
-export default interface CategoryView {
-  Id: number,
-  Name: string,
-}
