--- Buildlease/Buildlease/ClientApp/src/components/Views/ProductView.tsx
@@ -1,10 +1,0 @@
-export default interface ProductView {
-  Id: number,
-  Name: string,
-  Attributes: ProductAttributeView[],
-  ImagePath: string,
-  TotalCount: number,
-  AvailableCount: number,  
-  Price: number | null,
-  AlreadyInCart: boolean,
-}
