+++ Buildlease/Buildlease/ClientApp/src/components/views/ProductView.tsx
@@ -1,0 +1,12 @@
+import ProductAttributeView from "./ProductAttributeView";
+
+export default interface ProductView {
+  Id: number,
+  Name: string,
+  Attributes: ProductAttributeView[],
+  ImagePath: string,
+  TotalCount: number,
+  AvailableCount: number,  
+  Price: number | null,
+  AlreadyInCart: boolean,
+}
